refactor(backend): moved test helper functions to `internal/testutil` To be able to have tests located near their corresponding files, moved helper functions MakeTestUser and RegisterTestUser to a shareable, internal/testutil package, so all test packages that interact with the database can use them. Moved the setCleanup function into the second return value of MakeTestUser rather than a standalone function; the caller needs to run t.Cleanup(cleanup) to make sure the created test user is removed when the test ends.

+++ go-server/models/user.go
@@ -9,7 +9,7 @@ import (
 type User struct {
 	bun.BaseModel `bun:"table:users"`
 
-	ID        int32     `bun:"id,pk,autoincrement" json:"id"`
+	ID        int64     `bun:"id,pk,autoincrement" json:"id"`
 	Username  string    `bun:"username,type:varchar(50),unique,notnull" json:"username"`
 	PWHash    string    `bun:"password_hash,type:varchar(255),notnull" json:"-"`
 	Email     string    `bun:"email,type:varchar(255),unique,notnull" json:"email"`
